Define limits to screen refresh
I noticed that when increasing or decreasing the FPS using the + and -
keys, at some moment that the terminal would not be able to render the
game without blinking.

To solve the problem, I set limits for the scene update delay.

Note: I believe that this increase in fps will occur as the player get
more energy blocks, but I thought it to set it with limits for better
development.