Dynamically calculate the number of scenes
Create function countfiles that count how many scene files there are in the given directory.

If zero is giving as nscenes argument (the number of scenes) to readscenes function, then readscenes calls countfiles to calculate the actual nscenes, and scene argument is allocated for the appropriate size.

playmovie function has a new argument, nscenes, as this value is no more accessible by the hardcoded macro.

N_INTRO_SCENES is removed.